+++ src/Config/AbstractConfig.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Yansongda\Pay\Config;
 
+use Yansongda\Artful\Exception\InvalidConfigException;
+use Yansongda\Supports\Str;
 use Yansongda\Supports\Traits\Accessable;
 use Yansongda\Supports\Traits\Arrayable;
 use Yansongda\Supports\Traits\Serializable;
@@ -31,4 +33,19 @@ abstract class AbstractConfig implements ProviderConfigInterface
     abstract public function getMode(): int;
 
     abstract protected function validateRequired(): void;
+
+    /**
+     * @throws InvalidConfigException 缺少必要配置参数
+     */
+    protected function validateNotEmpty(array $props, int $exceptionCode, string $messagePrefix): void
+    {
+        foreach ($props as $prop) {
+            if (empty($this->{$prop})) {
+                throw new InvalidConfigException(
+                    $exceptionCode,
+                    $messagePrefix.' -- ['.Str::snake($prop).']'
+                );
+            }
+        }
+    }
 }